Feat: Update CloudFunction to call AppEngine API
This PR updates the cloudfunction to get all ShopDetails from AppEngine API instead of making three network calls. Earlier the API was making it's own network calls. Now, it calls the AppEngine API which makes those network calls on cloudFunction's behalf. 

This has been done to avoid cold starts which was causing a 3 second delay before the database was ready to be used. Since the AppEngine database connection is always alive, this solves the problem of cold start reducing the latency to < 500 ms from 3000 ms